chore: move prompts to a separate package (#1592)
